csv_options {xfun}	R Documentation

Parse comma-separated chunk options

Description

For knitr and R Markdown documents, code chunk options can be written using the comma-separated syntax (e.g., ⁠opt1=value1, opt2=value2⁠). This function parses these options and returns a list. If an option is not named, it will be treated as the chunk label.

Usage

csv_options(x)

Arguments

x
The chunk options as a string.

Value

A list of chunk options.

Examples

xfun::csv_options("foo, eval=TRUE, fig.width=5, echo=if (TRUE) FALSE")

[Package xfun version 0.47 Index]