dir_create {xfun}	R Documentation

Create a directory recursively by default

Description

First check if a directory exists. If it does, return TRUE, otherwise create it with dir.create(recursive = TRUE) by default.

Usage

dir_create(x, recursive = TRUE, ...)

Arguments

x
A path name.

recursive
Whether to create all directory components in the path.

...
Other arguments to be passed to dir.create().

Value

A logical value indicating if the directory either exists or is successfully created.

[Package xfun version 0.40 Index]