dir_exists {xfun}	R Documentation

Test the existence of files and directories

Description

These are wrapper functions of utils::file_test() to test the existence of directories and files. Note that file_exists() only tests files but not directories, which is the main difference between file.exists() in base R. If you use are using the R version 3.2.0 or above, dir_exists() is the same as dir.exists() in base R.

Usage

dir_exists(x)

file_exists(x)

Arguments

x
A vector of paths.

Value

A logical vector.

[Package xfun version 0.29 Index]