isFALSE {xfun}	R Documentation

Test if an object is FALSE

Description

For R versions lower than 3.5.0, this function is a simple abbreviation of identical(x, FALSE). For higher R versions, this function calls base::isFALSE().

Usage

isFALSE(x)

Arguments

x
An R object.

Note

This function will be deprecated in the future. We recommend that you use base::isFALSE() instead. If you have to support R versions lower than 3.5.0, you may use identical(x, FALSE), but please note that it is not equivalent to base::isFALSE().

Examples

library(xfun)
isFALSE(TRUE)  # false
isFALSE(FALSE)  # true
isFALSE(c(FALSE, FALSE))  # false

[Package xfun version 0.40 Index]