md5 {xfun}	R Documentation

Calculate the MD5 checksums of R objects

Description

Serialize an object to a temporary file, calculate the checksum via tools::md5sum(), and delete the file.

Usage

md5(...)

Arguments

...
Any number of R objects.

Value

A character vector of the checksums of objects passed to md5(). If the arguments are named, the results will also be named.

Examples

x1 = 1
x2 = 1:10
x3 = seq(1, 10)
x4 = iris
x5 = paste
(m = xfun::md5(x1, x2, x3, x4, x5))
stopifnot(m[2] == m[3])  # x2 and x3 should be identical

xfun::md5(x1 = x1, x2 = x2)  # named arguments

[Package xfun version 0.47 Index]