msg_cat {xfun}	R Documentation

Generate a message with cat()

Description

This function is similar to message(), and the difference is that msg_cat() uses cat() to write out the message, which is sent to stdout() instead of stderr(). The message can be suppressed by suppressMessages().

Usage

msg_cat(...)

Arguments

...
Character strings of messages, which will be concatenated into one string via paste(c(...), collapse = '').

Value

Invisible NULL, with the side-effect of printing the message.

Note

By default, a newline will not be appended to the message. If you need a newline, you have to explicitly add it to the message (see ‘Examples’).

See Also

This function was inspired by rlang::inform().

Examples

{
    # a message without a newline at the end
    xfun::msg_cat("Hello world!")
    # add a newline at the end
    xfun::msg_cat(" This message appears right after the previous one.\n")
}
suppressMessages(xfun::msg_cat("Hello world!"))

[Package xfun version 0.40 Index]