proc_kill {xfun}	R Documentation

Kill a process and (optionally) all its child processes

Description

Run the command taskkill /f /pid on Windows and kill on Unix, respectively, to kill a process.

Usage

proc_kill(pid, recursive = TRUE, ...)

Arguments

pid
The process ID.

recursive
Whether to kill the child processes of the process.

...
Arguments to be passed to system2() to run the command to kill the process.

Value

The status code returned from system2().

[Package xfun version 0.40 Index]