retry {xfun}	R Documentation

Retry calling a function for a number of times

Description

If the function returns an error, retry it for the specified number of times, with a pause between attempts.

Usage

retry(fun, ..., .times = 3, .pause = 5)

Arguments

fun
A function.

...
Arguments to be passed to the function.

.times
The number of times.

.pause
The number of seconds to wait before the next attempt.

Details

One application of this function is to download a web resource. Since the download might fail sometimes, you may want to retry it for a few more times.

Examples

# read the Github releases info of the repo yihui/xfun
xfun::retry(xfun::github_releases, "yihui/xfun")

[Package xfun version 0.40 Index]