rstudio_type {xfun}	R Documentation

Type a character vector into the RStudio source editor

Description

Use the rstudioapi package to insert characters one by one into the RStudio source editor, as if they were typed by a human.

Usage

rstudio_type(x, pause = function() 0.1, mistake = 0, save = 0)

Arguments

x
A character vector.

pause
A function to return a number in seconds to pause after typing each character.

mistake
The probability of making random mistakes when typing the next character. A random mistake is a random string typed into the editor and deleted immediately.

save
The probability of saving the document after typing each character. Note that If a document is not opened from a file, it will never be saved.

Examples

library(xfun)
if (loadable("rstudioapi") && rstudioapi::isAvailable()) {
    rstudio_type("Hello, RStudio! xfun::rstudio_type() looks pretty cool!",
        pause = function() runif(1, 0, 0.5), mistake = 0.1)
}

[Package xfun version 0.40 Index]