same_path {xfun}	R Documentation

Test if two paths are the same after they are normalized

Description

Compare two paths after normalizing them with the same separator (/).

Usage

same_path(p1, p2, ...)

Arguments

p1, p2
Two vectors of paths.

...
Arguments to be passed to normalize_path().

Examples

library(xfun)
same_path("~/foo", file.path(Sys.getenv("HOME"), "foo"))

[Package xfun version 0.40 Index]