set_envvar {xfun}	R Documentation

Set environment variables

Description

Set environment variables from a named character vector, and return the old values of the variables, so they could be restored later.

Usage

set_envvar(vars)

Arguments

vars
A named character vector of the form c(VARIABLE = VALUE). If any value is NA, this function will try to unset the variable.

Details

The motivation of this function is that Sys.setenv() does not return the old values of the environment variables, so it is not straightforward to restore the variables later.

Value

Old values of the variables (if not set, NA).

Examples

vars = xfun::set_envvar(c(FOO = "1234"))
Sys.getenv("FOO")
xfun::set_envvar(vars)
Sys.getenv("FOO")

[Package xfun version 0.40 Index]