submit_cran {xfun}	R Documentation

Submit a source package to CRAN

Description

Build a source package and submit it to CRAN with the curl package.

Usage

submit_cran(file = pkg_build(), comment = "")

Arguments

file
The path to the source package tarball. By default, the current working directory is treated as the package root directory, and automatically built into a tarball, which is deleted after submission. This means you should run xfun::submit_cran() in the root directory of a package project, unless you want to pass a path explicitly to the file argument.

comment
Submission comments for CRAN. By default, if a file ‘cran-comments.md’ exists, its content will be read and used as the comment.

See Also

devtools::submit_cran() does the same job, with a few more dependencies in addition to curl (such as cli); xfun::submit_cran() only depends on curl.

[Package xfun version 0.40 Index]