upload_ftp {xfun}	R Documentation

Upload to an FTP server via curl

Description

Run the command curl -T file server to upload a file to an FTP server. These functions require the system package (not the R package) curl to be installed (which should be available on macOS by default). The function upload_win_builder() uses upload_ftp() to upload packages to the win-builder server.

Usage


upload_ftp(file, server, dir = "")

upload_win_builder(file, version = c("R-devel", "R-release", "R-oldrelease"), 
    server = "ftp://win-builder.r-project.org/")

Arguments

file
Path to a local file.

server
The address of the FTP server.

dir
The remote directory to which the file should be uploaded.

version
The R version(s) on win-builder.

Details

These functions were written mainly to save package developers the trouble of going to the win-builder web page and uploading packages there manually. You may also consider using devtools::check_win_*, which currently only allows you to upload a package to one folder on win-builder each time, and xfun::upload_win_builder() uploads to all three folders, which is more likely to be what you need.

Value

Status code returned from system2.

[Package xfun version 0.6 Index]