broken_packages {xfun}	R Documentation

Find out broken packages and reinstall them

Description

If a package is broken (i.e., not loadable()), reinstall it.

Usage

broken_packages(reinstall = TRUE)

Arguments

reinstall
Whether to reinstall the broken packages, or only list their names.

Details

Installed R packages could be broken for several reasons. One common reason is that you have upgraded R to a newer x.y version, e.g., from 4.0.5 to 4.1.0, in which case you need to reinstall previously installed packages.

Value

A character vector of names of broken package.

[Package xfun version 0.29 Index]