cache_rds {xfun}	R Documentation

Cache the value of an R expression to an RDS file

Description

Save the value of an expression to a cache file (of the RDS format). Next time the value is loaded from the file if it exists. To invalidate the cache, you can either delete the file, or use the argument rerun = TRUE.

Usage

cache_rds(expr = {
}, rerun = FALSE, file = "cache.rds", dir = "cache/", ...)

Arguments

expr
An R expression.

rerun
Whether to delete the RDS file, rerun the expression, and save the result again (i.e., invalidate the cache).

file
The cache filename under the directory specified by the dir argument. If not specified and this function is called inside a code chunk of a knitr document (e.g., an R Markdown document), the filename is the current chunk label with the extension ‘.rds’.

dir
The path of the RDS file is determined by paste0(dir, file). If not specified and the knitr package is available, the default value of dir is the knitr chunk option cache.path (so if you are compiling a knitr document, you do not need to provide this dir argument explicitly), otherwise the default is ‘cache/’. If you do not want to provide a dir but simply a valid path to the file argument, you may use dir = "".

...
Other arguments to be passed to saveRDS().

Value

If the cache file does not exist, run the expression and save the result to the file, otherwise read the cache file and return the value.

Note

When this function is called in a code chunk of a knitr document, you may not want to provide the filename or directory of the cache file, because they have reasonable defaults.

Examples

f = tempfile()  # the cache file
compute = function(...) {
    res = xfun::cache_rds({
        Sys.sleep(2)
        1:10
    }, file = f, dir = "", ...)
    res
}
compute()  # takes two seconds
compute()  # returns 1:10 immediately
compute()  # fast again
compute(rerun = TRUE)  # two seconds to rerun
compute()
file.remove(f)

[Package xfun version 0.13 Index]