del_empty_dir {xfun}	R Documentation

Delete an empty directory

Description

Use list.file() to check if there are any files or subdirectories under a directory. If not, delete this empty directory.

Usage

del_empty_dir(dir)

Arguments

dir
Path to a directory. If NULL or the directory does not exist, no action will be performed.

[Package xfun version 0.29 Index]