download_cache {xfun}	R Documentation

Download a file from a URL and cache it on disk

Description

This object provides methods to download files and cache them on disk.

Usage

download_cache

Format

A list of methods:

$get(url, type, handler) downloads a URL, caches it, and returns the file content according to the value of type (possible values: "text" means the text content; "base64" means the base64 encoded data; "raw" means the raw binary content; "auto" is the default and means the type is determined by the content type in the URL headers). Optionally a handler function can be applied to the content.

$list() gives the list of cache files.

$summary() gives a summary of existing cache files.

$remove(url, type) removes a single cache file.

$purge() deletes all cache files.

Examples

# the first time it may take a few seconds
x1 = xfun::download_cache$get("https://www.r-project.org/")
head(x1)

# now you can get the cached content
x2 = xfun::download_cache$get("https://www.r-project.org/")
identical(x1, x2)  # TRUE

# a binary file
x3 = xfun::download_cache$get("https://yihui.org/images/logo.png", "raw")
length(x3)

# show a summary
xfun::download_cache$summary()
# remove a specific cache file
xfun::download_cache$remove("https://yihui.org/images/logo.png", "raw")
# remove all cache files
xfun::download_cache$purge()

[Package xfun version 0.38 Index]