github_releases {xfun}	R Documentation

Get the tags of Github releases of a repository

Description

Use the Github API (github_api()) to obtain the tags of the releases.

Usage

github_releases(
  repo,
  tag = "",
  pattern = "v[0-9.]+",
  use_jsonlite = loadable("jsonlite")
)

Arguments

repo
The repository name of the form user/repo, e.g., "yihui/xfun".

tag
A tag as a character string. If provided, it will be returned if the tag exists. If tag = "latest", the tag of the latest release is returned.

pattern
A regular expression to match the tags.

use_jsonlite
Whether to use jsonlite to parse the releases info.

Value

A character vector of (GIT) tags.

Examples

xfun::github_releases("yihui/xfun")
xfun::github_releases("gohugoio/hugo")

[Package xfun version 0.29 Index]