install_dir {xfun}	R Documentation

Install a source package from a directory

Description

Run R CMD build to build a tarball from a source directory, and run R CMD INSTALL to install it.

Usage

install_dir(pkg, build = TRUE, build_opts = NULL, install_opts = NULL)

Arguments

pkg
The package source directory.

build
Whether to build a tarball from the source directory. If FALSE, run R CMD INSTALL on the directory directly (note that vignettes will not be automatically built).

build_opts
The options for R CMD build.

install_opts
The options for R CMD INSTALL.

Value

Invisible status from R CMD INSTALL.

[Package xfun version 0.29 Index]