install_github {xfun}	R Documentation

An alias of remotes::install_github()

Description

This alias is to make autocomplete faster via xfun::install_github, because most remotes::install_* functions are never what I want. I only use install_github and it is inconvenient to autocomplete it, e.g. install_git always comes before install_github, but I never use it. In RStudio, I only need to type xfun::ig to get xfun::install_github.

Usage

install_github(...)

Arguments

...
Arguments to be passed to remotes::install_github().

[Package xfun version 0.29 Index]