is_blank {xfun}	R Documentation

Test if a character vector consists of blank strings

Description

Return TRUE if all elements of a character vector are blank (white spaces or empty strings).

Usage

is_blank(x)

Arguments

x
A character vector.

Value

TRUE if the input is blank, or FALSE otherwise.

Examples

xfun::is_blank("")
xfun::is_blank("abc")
xfun::is_blank(c("", "  ", "\n\t"))
xfun::is_blank(c("", " ", "abc"))

[Package xfun version 0.34 Index]