is_windows {xfun}	R Documentation

Test for types of operating systems

Description

Functions based on .Platform$OS.type and Sys.info() to test if the current operating system is Windows, macOS, Unix, or Linux.

Usage

is_windows()

is_unix()

is_macos()

is_linux()

Examples

library(xfun)
# only one of the following statements should be true
is_windows()
is_unix() && is_macos()
is_linux()

[Package xfun version 0.29 Index]