prose_index {xfun}	R Documentation

Find the indices of lines in Markdown that are prose (not code blocks)

Description

Filter out the indices of lines between code block fences such as ``` (could be three or four or more backticks).

Usage

prose_index(x, warn = TRUE)

Arguments

x
A character vector of text in Markdown.

warn
Whether to emit a warning when code fences are not balanced.

Value

An integer vector of indices of lines that are prose in Markdown.

Note

If the code fences are not balanced (e.g., a starting fence without an ending fence), this function will treat all lines as prose.

Examples

library(xfun)
prose_index(c("a", "```", "b", "```", "c"))
prose_index(c("a", "````", "```r", "1+1", "```", "````", "c"))

[Package xfun version 0.29 Index]