read_bin {xfun}	R Documentation

Read all records of a binary file as a raw vector by default

Description

This is a wrapper function of readBin() with default arguments what = "raw" and n = file.size(file), which means it will read the full content of a binary file as a raw vector by default.

Usage

read_bin(file, what = "raw", n = file.info(file)$size, ...)

Arguments

file, what, n, ...
Arguments to be passed to readBin().

Value

A vector returned from readBin().

Examples

f = tempfile()
cat("abc", file = f)
xfun::read_bin(f)
unlink(f)

[Package xfun version 0.29 Index]