read_utf8 {xfun}	R Documentation

Read / write files encoded in UTF-8

Description

Read or write files, assuming they are encoded in UTF-8. read_utf8() is roughly readLines(encoding = 'UTF-8') (a warning will be issued if non-UTF8 lines are found), and write_utf8() calls writeLines(enc2utf8(text), useBytes = TRUE).

Usage

read_utf8(con, error = FALSE)

write_utf8(text, con, ...)

append_utf8(text, con, sort = TRUE)

append_unique(text, con, sort = function(x) base::sort(unique(x)))

Arguments

con
A connection or a file path.

error
Whether to signal an error when non-UTF8 characters are detected (if FALSE, only a warning message is issued).

text
A character vector (will be converted to UTF-8 via enc2utf8()).

...
Other arguments passed to writeLines() (except useBytes, which is TRUE in write_utf8()).

sort
Logical (FALSE means not to sort the content) or a function to sort the content; TRUE is equivalent to base::sort.

Details

The function append_utf8() appends UTF-8 content to a file or connection based on read_utf8() and write_utf8(), and optionally sort the content. The function append_unique() appends unique lines to a file or connection.

[Package xfun version 0.29 Index]